Moved the ActivityPub to the Newsfeed directory.
